Start breaking apart Linkbot::Plugin
Linkbot::Plugin will remain the base class for all plugins.
Plus, .has_permission? was flattened and is readable

Linkbot::Message will become the handler/object that is passed around
Linkbot::Periodic (or something) will be the periodic runner, maybe?
